Add user input and vizualization module
The user can know enter how many words should be pulled from Random
Word API. The input is validated if it is a number between 5 and 20 and
is an integer.

Vizualization class for now has only one function that displays the
pulled words to user in a more readable way.