Partially fix bug with too many requests
The number of search querys in MusicBrainz API is limitted and it
returns 503 if the limit is exceeded. The user now can see a specific
message appearing and can try again. I also added 1.1 s delay after 4
querys to make the error appear less often. This caused the increased
time in waiting for search results.